+++ src/app/auth/services/index.ts
@@ -1,3 +1,2 @@
 export * from './auth.service';
-export * from './language.service';
 export * from './user.service';